incremental user prompts
